+++ testData/editor/refactoring/RenameVariableTestCase/basicRename/source.bash
@@ -1,0 +1,6 @@
+. target.bash
+
+$<caret>a
+echo $a
+eval "$a"
+eval '$a'